+++ .gitignore
@@ -1,4 +1,3 @@
 /backend/node_modules/
-/backend_new/node_modules/
-/backend_new/.env
+/backend/.env
 /frontend/node_modules/
