--- backend/data/posts.json
@@ -1,1 +1,0 @@
-[{"text":"bump","threadId":"1","id":1},{"text":"bump","threadId":"1","id":2},{"text":"bump","threadId":"3","id":3}]
