--- backend/data/threads.json
@@ -1,1 +1,0 @@
-[{"header":"тред 1","text":"йцйцуйцуй цуйцуй цуйц у","id":1,"timeValue":1701426333426,"timeCreation":1701426309902,"postCount":2},{"header":"тред 2","text":"фывфыцфсвфцвфцвфц","id":2,"timeValue":1701426315181,"timeCreation":1701426315181,"postCount":0},{"header":"тред 3","text":"фвыфыв фыв фы фывфывфыв фыв фывфывфы вф","id":3,"timeValue":1701426342188,"timeCreation":1701426321998,"postCount":1}]
