--- backend/framework/middlewares/parseUrl.js
@@ -1,8 +1,0 @@
-module.exports = (baseUrl) => (req, res) => {
-    const parsedUrl = new URL(req.url, baseUrl)
-    const params = {}
-    parsedUrl.searchParams.forEach((value, key) => params[key] = value)
-
-    req.pathname = parsedUrl.pathname
-    req.params = params
-}
