+++ backend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "imageboard-backend",
+  "name": "backend_new",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -9,9 +9,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "nodemon": "^3.0.1"
-  },
-  "nodemonConfig": {
-    "ignore": ["data/*"]
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "express-validator": "^7.0.1",
+    "moment": "^2.29.4",
+    "nodemon": "^3.0.2",
+    "pg": "^8.11.3"
   }
 }
