+++ backend/services/ThreadHidingService.js
