--- backend/src/post-router.js
@@ -1,8 +1,0 @@
-const Router = require('../framework/router')
-const controller = require('./post-controller')
-const router = new Router()
-
-router.get('/posts', controller.getPosts)
-router.post('/posts', controller.addPost)
-
-module.exports = router
