--- backend_new/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "backend_new",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "dev": "nodemon index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "express-validator": "^7.0.1",
-    "moment": "^2.29.4",
-    "nodemon": "^3.0.2",
-    "pg": "^8.11.3"
-  }
-}
